+++ cmd/zbackupd/main.go
@@ -36,7 +36,7 @@ import (
 )
 
 const (
-	version     = "0.1.0"
+	version     = "0.1.1"
 	serviceName = "zbackup"
 	binaryPath  = "/usr/bin/zbackupd"
 	dataPath    = "/DATA/AppData/zbackup"